updated tailwindcss to 1.0
