+++ app/ui/download.js
@@ -22,7 +22,9 @@ function downloading(state, emit) {
     <div
       class="flex flex-col w-full h-full items-center md:justify-center md:-mt-8"
     >
-      <h1 class="mb-4">${state.translate('downloadingTitle')}</h1>
+      <h1 class="text-3xl font-bold mb-4">
+        ${state.translate('downloadingTitle')}
+      </h1>
       ${archiveTile.downloading(state, emit)}
     </div>
   `;
@@ -36,7 +38,9 @@ function preview(state, emit) {
     <div
       class="flex flex-col w-full max-w-md h-full mx-auto items-center justify-center"
     >
-      <h1 class="mb-4">${state.translate('downloadTitle')}</h1>
+      <h1 class="text-3xl font-bold mb-4">
+        ${state.translate('downloadTitle')}
+      </h1>
       <p class="w-full text-grey-darkest text-center leading-normal">
         ${state.translate('downloadDescription')}
       </p>
