+++ app/ui/downloadPassword.js
@@ -8,7 +8,9 @@ module.exports = function(state, emit) {
     <div
       class="h-full w-full flex flex-col items-center justify-center bg-white py-8 max-w-md mx-auto"
     >
-      <h1 class="mb-4">${state.translate('downloadTitle')}</h1>
+      <h1 class="text-3xl font-bold mb-4">
+        ${state.translate('downloadTitle')}
+      </h1>
       <p class="w-full mb-4 text-center text-grey-darkest leading-normal">
         ${state.translate('downloadDescription')}
       </p>
