+++ app/ui/error.js
@@ -9,7 +9,7 @@ module.exports = function(state, emit) {
       <section
         class="flex flex-col items-center justify-center h-full w-full p-6 md:p-8 overflow-hidden md:rounded-xl md:shadow-big"
       >
-        <h1 class="text-center font-bold my-2">
+        <h1 class="text-center text-3xl font-bold my-2">
           ${state.translate('errorPageHeader')}
         </h1>
         <img class="my-12 h-48" src="${assets.get('error.svg')}" />
