+++ app/ui/surveyDialog.js
@@ -13,7 +13,7 @@ module.exports = function() {
       <send-survey-dialog
         class="flex flex-col items-center text-center p-4 max-w-sm m-auto"
       >
-        <h1 class="font-bold my-4">
+        <h1 class="text-3xl font-bold my-4">
           Tell us what you think.
         </h1>
         <p class="font-normal leading-normal text-grey-darkest px-4">
@@ -21,7 +21,7 @@ module.exports = function() {
           it better.
         </p>
         <a
-          class="btn rounded-lg w-full flex-no-shrink focus:outline my-5"
+          class="btn rounded-lg w-full flex-shrink-0 focus:outline my-5"
           onclick="${() => emit('closeModal')}"
           title="Give feedback"
           href="${surveyUrl}"
